Use the OMDb API to show movie information
In order to make use of the api functionality, the user needs to provide
their own valid api key. This is because the site is static, so
otherwise my api key would be embedded in the source code which allows
for malicious usage.

The plan is to maybe store the api key in a cookie so the user doesn't
have to enter it every time.